vite config ts=>js and NewUserForm
