+++ package.json
@@ -36,6 +36,7 @@
     "isomorphic-fetch": "3.0.0",
     "jsdom": "20.0.1",
     "msw": "0.47.4",
+    "noop-ts": "1.0.4",
     "prettier": "2.7.1",
     "terser": "5.15.1",
     "typescript": "4.8.4",
